Fix typo/bug in generating full-volume spin-color-diagonal noise.
The bug makes it so that rather than generating Nsrc independent sources
the module generates one source and then copies it Nsrc times.

I've been told that there is currently a project to rewrite this section
and this bug is known; however, it seems prescient not to leave a bug
like this in the main repository, especially since the fix is so simple.