Bug in Bilinear fixed, output extended, bug in Momentum fixed
The Bilinear module changed the provided PropagatorFields by phasing
them. I added auxillary fields such that qIn and qOut are unchanged
and can be provided as input to other modules without any problems.
I also added a volume normalization which seems to be consistent with
Max's tests in the free theory. In addition to that I added
'ExternalLegs' as 16th and 17th entries in the output. This could make
the Module ExternalLeg redundant. I slightly modified the Metadata class
to include a string description in order to make the output consistent.

I also set the PropagatorField explicitly to zero before assigning the
momentum in MSource/Momentum.